mqtt-api/requirements.txt
paho-mqtt~=1.4.0
kafka-python~=1.4.6
fastavro~=0.21.20
